services/training/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
gunicorn==21.2.0
google-cloud-run
python-multipart
jinja2
google-cloud-storage==3.1.1